+++ src/main/java/me/drex/vanish/config/VanishConfig.java
@@ -15,4 +15,7 @@ public class VanishConfig {
     @Comment("Prevents vanished players from using chat")
     public boolean disableChat = true;
 
+    @Comment("Hide vanished players from entities, prevents hostile entities from targeting players, and more")
+    public boolean hideFromEntities = true;
+
 }