+++ src/main/java/me/drex/vanish/mixin/LivingEntityMixin.java
@@ -3,11 +3,14 @@ package me.drex.vanish.mixin;
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import me.drex.vanish.api.VanishAPI;
+import me.drex.vanish.config.ConfigManager;
 import net.minecraft.core.particles.ParticleOptions;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.LivingEntity;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(LivingEntity.class)
 public abstract class LivingEntityMixin {
@@ -24,6 +27,13 @@ public abstract class LivingEntityMixin {
             return original.call(level, particleOptions, x, y, z, count, dx, dy, dz, speed);
         }
         return 0;
+    }
+
+    @Inject(method = "canBeSeenByAnyone", at = @At("HEAD"), cancellable = true)
+    public void vanish_hideFromEntities(CallbackInfoReturnable<Boolean> cir) {
+        if (ConfigManager.INSTANCE.vanish().hideFromEntities && VanishAPI.isVanished((LivingEntity) (Object) this)) {
+            cir.setReturnValue(false);
+        }
     }
 
 }